+++ src/auth/response-cache.ts
@@ -19,6 +19,10 @@ export class WebApiResponseCache implements ResponseCache {
   private cache: Cache | null = null;
   private async getCache() {
     if (!this.cache) {
+      if (typeof caches === "undefined") {
+        return undefined;
+      }
+
       return (this.cache = await caches.open(CERT_CACHE_KEY));
     }
 
@@ -26,12 +30,12 @@ export class WebApiResponseCache implements ResponseCache {
   }
   async put(url: URL, response: Response): Promise<void> {
     const cache = await this.getCache();
-    await cache.put(url, response);
+    await cache?.put(url, response);
   }
 
   async get(url: URL): Promise<Response | undefined> {
     const cache = await this.getCache();
 
-    return cache.match(url);
+    return cache?.match(url);
   }
 }